Moved CourtCaseList to seperate component
